# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HireITian Job Board API",
        "description": "Complete API documentation for HireITian - A job board platform connecting employers and candidates",
        "termsOfService": "https://hireitian.com/terms",
        "contact": {
            "name": "HireITian Support",
            "url": "https://hireitian.com",
            "email": "support@hireitian.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.hireitian.com/api",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/v1/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Search approved jobs",
                "description": "Search and filter through available job listings",
                "operationId": "searchJobs",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query by job title",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "work_type",
                        "in": "query",
                        "description": "Filter by work type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "employment_type",
                        "in": "query",
                        "description": "Filter by employment type",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "experience_level",
                        "in": "query",
                        "description": "Filter by experience level",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Filter by category ID",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salary_min",
                        "in": "query",
                        "description": "Minimum salary filter",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "salary_max",
                        "in": "query",
                        "description": "Maximum salary filter",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Jobs retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": []
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/jobs/{slug}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details",
                "description": "Retrieve detailed information about a specific job",
                "operationId": "getJobBySlug",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum bearer token. Use: Bearer {token}",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}